Download raw prescription data for specified county (by state and county names) into R

Data from from non-contiguous states not yet processed and available. This raw data only includes data from pharmacies and practitioners and the drugs Hydrocodne and Oxycodone.

county_raw(county = "Mingo", state = "WV", key = "WaPo")

Arguments

county

Filter the data to only this county (e.g. 'Mingo')

state

Filter the data to county within this state (e.g. 'WV')

key

Key needed to make query successful

See also

Examples

# \donttest{
library(arcos)

mingo_wv <- county_raw(county = "Mingo", state="WV", key="WaPo")

head(mingo_wv)
#>   REPORTER_DEA_NO REPORTER_BUS_ACT               REPORTER_NAME
#> 1       PH0035964      DISTRIBUTOR H D SMITH WHOLESALE DRUG CO
#> 2       PH0035964      DISTRIBUTOR H D SMITH WHOLESALE DRUG CO
#> 3       PH0035964      DISTRIBUTOR H D SMITH WHOLESALE DRUG CO
#> 4       PH0035964      DISTRIBUTOR H D SMITH WHOLESALE DRUG CO
#> 5       PH0035964      DISTRIBUTOR H D SMITH WHOLESALE DRUG CO
#> 6       RD0277409      DISTRIBUTOR MASTERS PHARMACEUTICAL, INC
#>    REPORTER_ADDL_CO_INFO    REPORTER_ADDRESS1 REPORTER_ADDRESS2 REPORTER_CITY
#> 1                   null   4650 INDUSTRIAL DR              null   SPRINGFIELD
#> 2                   null   4650 INDUSTRIAL DR              null   SPRINGFIELD
#> 3                   null   4650 INDUSTRIAL DR              null   SPRINGFIELD
#> 4                   null   4650 INDUSTRIAL DR              null   SPRINGFIELD
#> 5                   null   4650 INDUSTRIAL DR              null   SPRINGFIELD
#> 6 DBA: RIVER CITY PHARMA 11930 KEMPER SPRINGS              null    CINCINNATI
#>   REPORTER_STATE REPORTER_ZIP REPORTER_COUNTY BUYER_DEA_NO   BUYER_BUS_ACT
#> 1             IL        62703        SANGAMON    BH6954401 RETAIL PHARMACY
#> 2             IL        62703        SANGAMON    BH6954401 RETAIL PHARMACY
#> 3             IL        62703        SANGAMON    BH6954401 RETAIL PHARMACY
#> 4             IL        62703        SANGAMON    BH6954401 RETAIL PHARMACY
#> 5             IL        62703        SANGAMON    BH6954401 RETAIL PHARMACY
#> 6             OH        45240        HAMILTON    BH6954401 RETAIL PHARMACY
#>                BUYER_NAME BUYER_ADDL_CO_INFO   BUYER_ADDRESS1 BUYER_ADDRESS2
#> 1 HURLEY DRUG COMPANY INC               null 210 LOGAN STREET           null
#> 2 HURLEY DRUG COMPANY INC               null 210 LOGAN STREET           null
#> 3 HURLEY DRUG COMPANY INC               null 210 LOGAN STREET           null
#> 4 HURLEY DRUG COMPANY INC               null 210 LOGAN STREET           null
#> 5 HURLEY DRUG COMPANY INC               null 210 LOGAN STREET           null
#> 6 HURLEY DRUG COMPANY INC               null 210 LOGAN STREET           null
#>   BUYER_CITY BUYER_STATE BUYER_ZIP BUYER_COUNTY TRANSACTION_CODE DRUG_CODE
#> 1 WILLIAMSON          WV     25661        MINGO                S      9193
#> 2 WILLIAMSON          WV     25661        MINGO                S      9193
#> 3 WILLIAMSON          WV     25661        MINGO                S      9193
#> 4 WILLIAMSON          WV     25661        MINGO                S      9193
#> 5 WILLIAMSON          WV     25661        MINGO                S      9143
#> 6 WILLIAMSON          WV     25661        MINGO                S      9193
#>        NDC_NO   DRUG_NAME QUANTITY UNIT ACTION_INDICATOR ORDER_FORM_NO
#> 1 00591050205 HYDROCODONE        3 null             null          null
#> 2 00591050205 HYDROCODONE        4 null             null          null
#> 3 52544053901 HYDROCODONE        2 null             null          null
#> 4 00591038705 HYDROCODONE        1 null             null          null
#> 5 52152021402   OXYCODONE        5 null             null     064960413
#> 6 00591050305 HYDROCODONE        2 null             null          null
#>   CORRECTION_NO STRENGTH TRANSACTION_DATE CALC_BASE_WT_IN_GM DOSAGE_UNIT
#> 1          null     null         01042007             6.8107        1500
#> 2          null     null         01112007             9.0810        2000
#> 3          null     null         01042007             1.2108         200
#> 4          null     null         01172007             2.2702         500
#> 5          null     null         01192007             6.7237         500
#> 6          null     0000         08082008             6.0540        1000
#>   TRANSACTION_ID                         Product_Name
#> 1          13693 HYDROCODONE BIT./ACET.,7.5MG & 650MG
#> 2          13713 HYDROCODONE BIT./ACET.,7.5MG & 650MG
#> 3          14058 NORCO HYDROCODO.BIT./ACET.,10MG & 32
#> 4           1644 HYDROCODONE BITARTRATE 7.5MG/ACETAMI
#> 5           8496  OXYCODONE HYDROCHLORIDE TABLETS USP
#> 6           1011 HYDROCODONE BITARTRATE 10MG/ACETAMIN
#>                           Ingredient_Name Measure MME_Conversion_Factor
#> 1 HYDROCODONE BITARTRATE HEMIPENTAHYDRATE     TAB                   1.0
#> 2 HYDROCODONE BITARTRATE HEMIPENTAHYDRATE     TAB                   1.0
#> 3 HYDROCODONE BITARTRATE HEMIPENTAHYDRATE     TAB                   1.0
#> 4 HYDROCODONE BITARTRATE HEMIPENTAHYDRATE     TAB                   1.0
#> 5                 OXYCODONE HYDROCHLORIDE     TAB                   1.5
#> 6 HYDROCODONE BITARTRATE HEMIPENTAHYDRATE     TAB                   1.0
#>   Combined_Labeler_Name Revised_Company_Name   Reporter_family dos_str
#> 1  Actavis Pharma, Inc.       Allergan, Inc.       H. D. Smith     7.5
#> 2  Actavis Pharma, Inc.       Allergan, Inc.       H. D. Smith     7.5
#> 3  Actavis Pharma, Inc.       Allergan, Inc.       H. D. Smith    10.0
#> 4  Actavis Pharma, Inc.       Allergan, Inc.       H. D. Smith     7.5
#> 5  Actavis Pharma, Inc.       Allergan, Inc.       H. D. Smith    15.0
#> 6  Actavis Pharma, Inc.       Allergan, Inc. River City Pharma    10.0
# }