Download raw prescription data for individual drugs and business type by county into R

Data from from non-contiguous states not yet processed and available.

drug_county_biz(
  drug = "Fentanyl",
  county = "Laurens",
  state = "GA",
  buyer_bus_act = "Chain Retail",
  key = "WaPo"
)

Arguments

drug
Filter the data to one of 14 available drugs in the database (e.g. 'OXYCODONE')

county
Filter the data to only this county (e.g. 'Mingo')

state
Filter the data to county within this state (e.g. 'WV')

buyer_bus_act
If provided, filter the data to one of businesses in BUYER_BUS_ACT (e.g. 'CHAIN PHARMACY', 'HOSP/CLINIC')

key
Key needed to make query successful

See also

Examples

# \donttest{
library(arcos)

bronx <- drug_county_biz(drug="Fentanyl", county = "Laurens", state = "GA",
                         buyer_bus_act = "Chain Retail", key="WaPo")

head(bronx)

#>   REPORTER_DEA_NO REPORTER_BUS_ACT        REPORTER_NAME    REPORTER_ADDRESS1
#> 1       PR0040357      DISTRIBUTOR MCKESSON CORPORATION DBA MCKESSON DRUG CO
#> 2       PR0040357      DISTRIBUTOR MCKESSON CORPORATION DBA MCKESSON DRUG CO
#> 3       PR0040357      DISTRIBUTOR MCKESSON CORPORATION DBA MCKESSON DRUG CO
#> 4       PR0040357      DISTRIBUTOR MCKESSON CORPORATION DBA MCKESSON DRUG CO
#> 5       PR0040357      DISTRIBUTOR MCKESSON CORPORATION DBA MCKESSON DRUG CO
#> 6       PR0040357      DISTRIBUTOR MCKESSON CORPORATION DBA MCKESSON DRUG CO
#>   REPORTER_ADDRESS2 REPORTER_CITY REPORTER_STATE REPORTER_ZIP REPORTER_COUNTY
#> 1 2975 EVERGREEN DR        DULUTH             GA        30096        GWINNETT
#> 2 2975 EVERGREEN DR        DULUTH             GA        30096        GWINNETT
#> 3 2975 EVERGREEN DR        DULUTH             GA        30096        GWINNETT
#> 4 2975 EVERGREEN DR        DULUTH             GA        30096        GWINNETT
#> 5 2975 EVERGREEN DR        DULUTH             GA        30096        GWINNETT
#> 6 2975 EVERGREEN DR        DULUTH             GA        30096        GWINNETT
#>   BUYER_DEA_NO   BUYER_BUS_ACT         BUYER_NAME     BUYER_ADDRESS1 BUYER_CITY
#> 1    AM5673860 RETAIL PHARMACY MAX BROWN PHARMACY 809 N JEFFERSON ST     DUBLIN
#> 2    AM5673860 RETAIL PHARMACY MAX BROWN PHARMACY 809 N JEFFERSON ST     DUBLIN
#> 3    AM5673860 RETAIL PHARMACY MAX BROWN PHARMACY 809 N JEFFERSON ST     DUBLIN
#> 4    AM5673860 RETAIL PHARMACY MAX BROWN PHARMACY 809 N JEFFERSON ST     DUBLIN
#> 5    AM5673860 RETAIL PHARMACY MAX BROWN PHARMACY 809 N JEFFERSON ST     DUBLIN
#> 6    AM5673860 RETAIL PHARMACY MAX BROWN PHARMACY 809 N JEFFERSON ST     DUBLIN
#>   BUYER_STATE BUYER_ZIP BUYER_COUNTY TRANSACTION_CODE DRUG_CODE      NDC_NO
#> 1          GA     31021      LAURENS                S      9801 00591321472
#> 2          GA     31021      LAURENS                S      9801 00406905076
#> 3          GA     31021      LAURENS                S      9801 00406907576
#> 4          GA     31021      LAURENS                S      9801 00409909422
#> 5          GA     31021      LAURENS                S      9801 10019003372
#> 6          GA     31021      LAURENS                S      9801 00591321372
#>   DRUG_NAME QUANTITY ORDER_FORM_NO TRANSACTION_DATE CALC_BASE_WT_IN_GM
#> 1  FENTANYL        6     11X000016         08252011             0.3000
#> 2  FENTANYL        4     11X000081         12022011             0.1100
#> 3  FENTANYL        4     12X000116         05302012             0.1650
#> 4  FENTANYL        3     12X000162         08022012             0.0075
#> 5  FENTANYL        1     12X000174         08212012             0.0025
#> 6  FENTANYL        6     12X000211         10172012             0.2250
#>   DOSAGE_UNIT TRANSACTION_ID                         Product_Name
#> 1          30     1108047712 FENTANYL 100MCG/HR TRANSDERMAL SYSTE
#> 2          20     1112030184 FENTANYL 72HR 50MCG/HR TDS (5.50MG T
#> 3          20     1205032191 FENTANYL 72HR 75MCG/HR TDS (8.25MG T
#> 4          NA     1208034161 FENTANYL 50MCG/ML INJECTABLE SOLUTIO
#> 5          NA     1208058628 FENTANYL  INJ., .05MG/ML (5ML AMPUL)
#> 6          30     1210048536 FENTANYL 75MCG/HR TRANSDERMAL SYSTEM
#>   Ingredient_Name Measure MME_Conversion_Factor         Combined_Labeler_Name
#> 1   FENTANYL BASE   PO/PA                   100          Actavis Pharma, Inc.
#> 2   FENTANYL BASE   PO/PA                   100                    SpecGx LLC
#> 3   FENTANYL BASE   PO/PA                   100                    SpecGx LLC
#> 4   FENTANYL BASE      ML                   100                 Hospira, Inc.
#> 5   FENTANYL BASE     AMP                   100 Baxter Healthcare Corporation
#> 6   FENTANYL BASE   PO/PA                   100          Actavis Pharma, Inc.
#>        Reporter_family   MME year BUYER_ADDL_CO_INFO BUYER_ADDRESS2 dos_str
#> 1 McKesson Corporation 30000 2011               <NA>           <NA>      NA
#> 2 McKesson Corporation 11000 2011               <NA>           <NA>      NA
#> 3 McKesson Corporation 16500 2012               <NA>           <NA>      NA
#> 4 McKesson Corporation   750 2012               <NA>           <NA>      NA
#> 5 McKesson Corporation   250 2012               <NA>           <NA>      NA
#> 6 McKesson Corporation 22500 2012               <NA>           <NA>      NA
#>   STRENGTH REPORTER_ADDL_CO_INFO ACTION_INDICATOR CORRECTION_NO
#> 1     <NA>                  <NA>             <NA>          <NA>
#> 2     <NA>                  <NA>             <NA>          <NA>
#> 3     <NA>                  <NA>             <NA>          <NA>
#> 4     <NA>                  <NA>             <NA>          <NA>
#> 5     <NA>                  <NA>             <NA>          <NA>
#> 6     <NA>                  <NA>             <NA>          <NA>

# }