Load data from file

Loads data from a file and creates an arcosTS object based on provided parameters. Data is loaded using data.table::fread function, thus all file formats recognised by fread are admissible, including gzip and bz2 compressions.

loadDataFromFile(
  fname,
  colPos = "x",
  colFrame = "f",
  colIDobj = "id",
  colIDcoll = NULL,
  colMeas = NULL,
  colMeasResc = NULL,
  colMeasBin = NULL,
  colBootIter = NULL,
  colRT = NULL,
  interVal = 1L,
  interType = c("fixed"),
  fromBin = FALSE,
  fromColl = FALSE,
  fromBoot = FALSE,
  ...
)

Arguments

fname

a string, file name path.

colPos

a vector, names of positional columns; default "x".

colFrame

a string, column name with frame numbers; default "f".

colIDobj

a string, column name with object IDs; default "id".

colIDcoll

a string, column name with IDs of collective events; default NULL.

colMeas

a string, column name of the measurement; default NULL.

colMeasResc

a string, column name of the rescaled measurement; default NULL.

colMeasBin

a string, column name of the binarised measurement; default NULL.

colBootIter

a string, column name with bootstrap iterations; default NULL.

colRT

a string, column name with real time; default NULL (currently unused).

interVal

a numeric with the interval length (currently unused).

interType

a string to designate whether the time series has fixed or variable intervals; possible values fixed or var (currently unused).

fromBin

logical, whether the output has been passed by binarisation, relevant for plotting of binarised activity; default FALSE.

fromColl

logical, whether the output comes from collective event identification, relevant for plotting of collective events; default FALSE.

fromBoot

logical, whether the output comes from bootstrapping, relevant for calculating stats of collective events; default FALSE.

...

additional parameters passed to data.table::fread.

Value

an arcosTS object.

Examples

cat("no examples")
#> no examples