Get total pills for each distributor in a state

Get total pills for each distributor in a state

total_distributors_state(state = "WV", key = "WaPo")

Arguments

state

Filter the data to county within this state (e.g. 'WV')

key

Key needed to make query successful

See also

Examples

# \donttest{
library(arcos)

wv <- total_distributors_state(state="WV", key="WaPo")

head(wv)
#>   buyer_state buyer_county        Reporter_family total_dosage_unit
#> 1          WV       CABELL AmerisourceBergen Drug          27564760
#> 2          WV      KANAWHA               Rite Aid          18987950
#> 3          WV      KANAWHA        Cardinal Health          18418060
#> 4          WV     HARRISON        Cardinal Health          15477110
#> 5          WV      HANCOCK        Cardinal Health          12907450
#> 6          WV        LOGAN   McKesson Corporation          12548310
#>   total_records
#> 1         51557
#> 2         29772
#> 3         52962
#> 4         36515
#> 5         31477
#> 6         15677
# }