Get total pills for each pharmacy in a county

Get total pills for each pharmacy in a county

total_pharmacies_county(county = "Mingo", state = "WV", key = "WaPo")

Arguments

county
Filter the data to only this county (e.g. 'Mingo')

state
Filter the data to county within this state (e.g. 'WV')

key
Key needed to make query successful

See also

Examples

# \donttest{
library(arcos)

mingo_wv <- total_pharmacies_county(county = "Mingo", state="WV", key="WaPo")

head(mingo_wv)

#>   buyer_state buyer_county buyer_dea_no               buyer_name buyer_city
#> 1          WV        MINGO    BS7437064               STROSNIDER     KERMIT
#> 2          WV        MINGO    BH6954401  HURLEY DRUG COMPANY INC WILLIAMSON
#> 3          WV        MINGO    FT0251227 TUG VALLEY PHARMACY, LLC WILLIAMSON
#> 4          WV        MINGO    FR0261684       RIVERSIDE PHARMACY    GILBERT
#> 5          WV        MINGO    AA8151728      ADKINS PHARMACY INC    GILBERT
#> 6          WV        MINGO    FG0153863         GILBERT PHARMACY    GILBERT
#>   total_dosage_unit total_records
#> 1          13168350          7691
#> 2           8890370         11138
#> 3           8827860          5390
#> 4           1780680          2249
#> 5           1576200          2491
#> 6           1403720          3138

# }