"Track collective events with bootstrapping randomisation"

Track collective events with bootstrapping randomisation

trackCollBoot(
  obj,
  nboot = 100L,
  ncores = 2L,
  method = c("shuffCoord", "randShiftMeas", "shuffMeasTrack", "shuffMeasFrame",
    "shuffBlockTrack"),
  eps = 1,
  minClSz = 1L,
  nPrev = 1L,
  epsPrev = NULL,
  colldurlim = c(1, Inf),
  colltotszlim = c(1, Inf),
  deb = FALSE
)

# S3 method for arcosTS
trackCollBoot(
  obj,
  nboot = 100L,
  ncores = 2L,
  method = c("shuffCoord", "randShiftMeas", "shuffMeasTrack", "shuffMeasFrame",
    "shuffBlockTrack"),
  eps = 1,
  minClSz = 1L,
  nPrev = 1L,
  epsPrev = NULL,
  colldurlim = c(1, Inf),
  colltotszlim = c(1, Inf),
  deb = FALSE
)

Arguments

obj

an arcosTS object.

nboot

an integer, number of bootstrap iterations; default 100.

ncores

an integer, number of parallel cores; default 2.

method

either of the five bootstrapping methods: shuffCoord', 'randShiftMeas', 'shuffMeasTrack', 'shuffMeasFrame', 'shuffBlockTrack'.

eps

a numeric, sets the search radius for spatial clustering with dbscan; default 1.

minClSz

an integer, minimum cluster size for dbscan; default 1L.

nPrev

an integer, number of previous frames to link; default 1L.

epsPrev

a float with the search radius for linking clusters between frames, default NULL means that epsPrev = eps.

colldurlim

a two-element vector with limits for filtering cluster duration; default c(1, Inf).

colltotszlim

a two-element vector with limits for filtering cluster size; default c(1, Inf).

deb

boolean, additional debug output; default FALSE.

Value

a data.table with filtered stats of collective events at every bootstrapping iteration

Details

Identifies collective events using one of five randomisation methods.

Examples

cat("No example\n")
#> No example